rewrite: refactor snapshot generator unit tests, move utility classes to separate files
